# api/docs/openapi.yaml
...
openapi: 3.0.0
info:
  title: Ninaivalaigal Agentic API
  description: AI-powered memory management and context-aware API
  version: 1.0.0
  contact:
    name: Ninaivalaigal Team
    url: https://github.com/Arunosaur/ninaivalaigal
servers:
  - url: http://localhost:13370
    description: Development server (API container)
  - url: http://localhost:8000
    description: Development server (direct)
  - url: https://api.ninaivalaigal.com
    description: Production server
paths:
...
  /contexts:
    get:
      summary: List user contexts
      tags: [Contexts]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of contexts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Context'
    post:
      summary: Create a new context
      tags: [Contexts]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                scope:
                  type: string
                  enum: [personal, team, organization, global]
                  default: personal
      responses:
        '201':
          description: Context created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Context'
  /contexts/{context_id}:
    get:
      summary: Get context details
      tags: [Contexts]
      security:
        - bearerAuth: []
      parameters:
        - name: context_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Context details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Context'
        '404':
          description: Context not found
    put:
      summary: Update context
      tags: [Contexts]
      security:
        - bearerAuth: []
      parameters:
        - name: context_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                is_active:
                  type: boolean
      responses:
        '200':
          description: Context updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Context'
    delete:
      summary: Delete context
      tags: [Contexts]
      security:
        - bearerAuth: []
      parameters:
        - name: context_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Context deleted successfully
        '404':
          description: Context not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Context:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        scope:
          type: string
          enum: [personal, team, organization, global]
        is_active:
          type: boolean
        owner_id:
          type: integer
        created_at:
          type: string
          format: date-time
        memory_count:
          type: integer
...
security:
  - bearerAuth: []
...